refactor: unify Geth types and fix EIP-8037/EIP-7702 test regressions

+++ src/Nethermind/Nethermind.Benchmark/State/ReadOnlySnapshotBundleBenchmark.cs
@@ -66,7 +66,7 @@ public class ReadOnlySnapshotBundleBenchmark
         // Track storage account ranges per snapshot for hit distribution
         List<(int AddressStart, int StorageCount, int SlotsPerAccount)> storageRanges = [];
 
-        for (int block = 0; block < SnapshotCount; block++)
+        for (ulong block = 0; block < SnapshotCount; block++)
         {
             int multiplier = block < 6 ? 16 : 1;
             int accountCount = 1000 * multiplier;
@@ -138,7 +138,7 @@ public class ReadOnlySnapshotBundleBenchmark
                 });
             }
 
-            scope.Commit(blockNumber: (ulong)(block + 1));
+            scope.Commit(blockNumber: block + 1);
 
             FlatSnapshot snapshot = commitTarget.LastSnapshot
                 ?? throw new InvalidOperationException(
@@ -146,7 +146,7 @@ public class ReadOnlySnapshotBundleBenchmark
             snapshot.TryAcquire();
             allSnapshots.Add(snapshot);
 
-            currentStateId = new StateId((ulong)(block + 1), scope.RootHash);
+            currentStateId = new StateId(block + 1, scope.RootHash);
             storageRanges.Add((totalAccountCount + 1, storageAccountCount, slotsPerStorageAccount));
             totalAccountCount += accountCount;
             totalStorageAccountCount += storageAccountCount;
