+++ src/Nethermind/Nethermind.Benchmark/Store/PatriciaTreeBenchmarks.cs
@@ -351,15 +351,15 @@ namespace Nethermind.Benchmarks.Store
                 Persist.EveryNBlock(2), NullLogManager.Instance);
             StateTree tempTree = new(trieStore, NullLogManager.Instance);
 
-            for (int i = 0; i < _largerEntryCount; i++)
+            for (ulong i = 0; i < _largerEntryCount; i++)
             {
                 if (i % _repeatedlyFactor == 0)
                 {
-                    using IBlockCommitter _ = trieStore.BeginBlockCommit((ulong)(i / _repeatedlyFactor));
+                    using IBlockCommitter _ = trieStore.BeginBlockCommit(i / _repeatedlyFactor);
                     tempTree.Commit();
                 }
 
-                (bool isWrite, Hash256 address, Account value) = _largerEntriesAccess[i];
+                (bool isWrite, Hash256 address, Account value) = _largerEntriesAccess[(int)i];
                 if (isWrite)
                 {
                     tempTree.Set(address, value);
